Add DateField type
